Things Have Gotten Good
So, I added stack to pull in some dependencies for making PNGs and stuff.
I've decided that since the animation tools don't really care what the `a` in
`Animation a` is, it can be a magic custom type.

Then I can just animate it, and render that outside the library.
Which is exatly what I've done.

Then the question was "How do I combine animations" and the answer there was
that it was nonsense to combine animations, they're too generic.
How do you combine to animations of a Float?

The only reason I think I can combine things is because I know there are
separate facets to my scene tha I can compare independently.
So I made an "Animator" which is just an `Animation (a -> a)`
Given a function that modifies another value over time, I can animate it over
time and get my value out.
That's allowed me to then start talking about combining things, since I can
start with a default value, and then only apply some modifications later.

The best part is that since Animators are just Animations, my pure Animation
combinators like `after` and `holdFor` don't have to care that they're
animators.
They still work!

Best thing ever.