Lenses
So I've taken what I had already, and then used the lenses to try to tighten up
some of the interactions.

I had to introduce a hack of sorts to get there, because I can't produce a Just
with a value inside it in one lens operation, so instead I have two animations.

The first produces a default object, and the next populates it with something.
Despite the fact that it's garbage, it's nearly elegant.

I'm also seeing a lot of familiar shapes in these abstractions, but I'm
resisting the urge to implement typeclasses yet.