Added Automatic Starting Point Tween Operator
Wow!
This is so much less hacky than I was afraid of!

So, basically, I've made it so that since you gave me a lens anyway, I can just
query the current value.

So that's what this new lens operator does.
It says "Tween this from wherever it is to this new location"

It means I don't have to duplicate my starting point all over the place, which
was super poop.

The other thing it means, though, is I don't have to keep track of where the
object ended up after all of the animation.
So I've added a new movement that moves alice from seconds 4 to 5.

Previously, as part of the animation at second 10 it would have shifted alice
back, then begun her animation from where I said she used to be, since I would
have forgotten to update her "starting value" with the new value after I moved
her.

Now, though, that function is applied in the stack, and then we go and query
the value to say "Ok, tween from where you are to where I want you" and it just
works!

And it doesn't have to interative, since all the animations are still
declarative.
I can render a single frame out of the middle without rendering everything up
until that point, or I can render all frames in parallel, or whatever.

So good!