Add Quint Ease and Hold Function
I've realized that how long something is held for is largely irrelevant,
because the animation ends up being longer for other reasons anyway.

But, I do need to keep some of my animations from running away, so that's where
hold comes in.
It just freezes the last moment forever, so when I don't care how long it
pauses at the end, I can just use this.