Add Ability to Pick Single Frame
While I'm positioning things it can help to be able to look at a particular
time, rather than waiting for the whole animation to render and play.

So, now I can do `make view TIME=5.2` and it shows me what I want!
Neato!

(Also it has time as an input instead of frame because it's meant to be roughly
framerate independent)