Change from picked.png to frameT13.png
When I changed from one FRAME to another on the cli, it wouldn't do anything
because it didn't see the code change, and I couldn't have it depend on the
env.

So it was easiest to just have it put the frame in a filename, that way I can
ask about different frames without them colliding.